improvements to trigram efficiency
improvments inherently involve a change in function signature to
allow the calling function to pass in the result slice and start
location as input parameters. This allows the caller to do all the
memory allocation.

This helps us as the Tokenize method can pre-compute the final
slice size before doing a single large allocation.

The new function is unexported, while the main Trigram method
remains exported to preserve the func signature. However, since
this method must still allocate its own memory, there is no
significant improvement here. What IS improved greater is the
way Tokenize uses the unexported trigram function.

There is also a removal of an unecessary if branch.